comments fixes
