+++ service/worker/batcher/workflow_v2.go
@@ -25,17 +25,12 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/google/uuid"
 	"go.uber.org/cadence"
 	"go.uber.org/cadence/activity"
 	"go.uber.org/cadence/workflow"
 	"golang.org/x/time/rate"
 
 	"github.com/uber/cadence/client/admin"
-	"github.com/uber/cadence/client/frontend"
-	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/log/tag"
-	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/types"
 )
 
@@ -58,17 +53,6 @@ type TuneSignal struct {
 	Concurrency int
 }
 
-// batchParamsV2 extends BatchParams with V2-specific fields.
-// It is used only by the V2 activity and is not part of the public API.
-type batchParamsV2 struct {
-	BatchParams
-
-	// Progress carries forward HeartBeatDetails from a cancelled activity
-	// so the next activity invocation can resume where the previous left off.
-	// nil means no prior progress (use heartbeat recovery or start fresh).
-	Progress *HeartBeatDetails
-}
-
 func init() {
 	workflow.RegisterWithOptions(BatchWorkflowV2, workflow.RegisterOptions{Name: BatchWFV2TypeName})
 	activity.RegisterWithOptions(batchActivityV2, activity.RegisterOptions{Name: batchActivityV2Name})
@@ -85,22 +69,22 @@ func BatchWorkflowV2(ctx workflow.Context, batchParams BatchParams) (HeartBeatDe
 	}
 
 	tuneCh := workflow.GetSignalChannel(ctx, SignalNameTune)
-	v2Params := batchParamsV2{BatchParams: batchParams}
+	params := batchParams
 
 	for {
 		retryPolicy := BatchActivityRetryPolicy
-		retryPolicy.MaximumAttempts = int32(v2Params.MaxActivityRetries)
+		retryPolicy.MaximumAttempts = int32(params.MaxActivityRetries)
 		actOpts := workflow.ActivityOptions{
 			ScheduleToStartTimeout: 5 * time.Minute,
 			StartToCloseTimeout:    InfiniteDuration,
-			HeartbeatTimeout:       v2Params.ActivityHeartBeatTimeout,
+			HeartbeatTimeout:       params.ActivityHeartBeatTimeout,
 			RetryPolicy:            &retryPolicy,
 			WaitForCancellation:    true,
 		}
 		actCtx, cancel := workflow.WithCancel(workflow.WithActivityOptions(ctx, actOpts))
 
 		var result HeartBeatDetails
-		future := workflow.ExecuteActivity(actCtx, batchActivityV2Name, v2Params)
+		future := workflow.ExecuteActivity(actCtx, batchActivityV2Name, params)
 
 		selector := workflow.NewSelector(ctx)
 		var actErr error
@@ -115,48 +99,38 @@ func BatchWorkflowV2(ctx workflow.Context, batchParams BatchParams) (HeartBeatDe
 			var sig TuneSignal
 			ch.Receive(ctx, &sig)
 			if sig.RPS > 0 {
-				v2Params.RPS = sig.RPS
+				params.RPS = sig.RPS
 			}
 			if sig.Concurrency > 0 {
-				v2Params.Concurrency = sig.Concurrency
+				params.Concurrency = sig.Concurrency
 			}
-			// Drain any additional pending signals.
-			for ch.ReceiveAsync(&sig) {
-				if sig.RPS > 0 {
-					v2Params.RPS = sig.RPS
-				}
-				if sig.Concurrency > 0 {
-					v2Params.Concurrency = sig.Concurrency
-				}
-			}
-			// Cancel the running activity so it returns with current progress.
-			cancel()
 		})
 
 		selector.Select(ctx)
 
 		if activityDone {
+			cancel()
 			return result, actErr
 		}
 
-		// Activity was cancelled by tune signal — wait for it to finish
-		// and extract progress from the CanceledError.
-		err := future.Get(ctx, nil)
-		cancel() // release resources
-
-		if err != nil {
-			if cadence.IsCanceledError(err) {
-				if ce, ok := err.(*cadence.CanceledError); ok {
-					var hbd HeartBeatDetails
-					if ce.Details(&hbd) == nil {
-						v2Params.Progress = &hbd
-					}
+		// Tune signal fired — cancel the activity and wait for it to finish.
+		cancel()
+		err := future.Get(ctx, &result)
+		if err == nil {
+			// Activity finished before the cancellation was delivered — return its result directly.
+			return result, nil
+		}
+		if cadence.IsCanceledError(err) {
+			if ce, ok := err.(*cadence.CanceledError); ok {
+				var hbd HeartBeatDetails
+				if ce.Details(&hbd) == nil {
+					params.Progress = &hbd
 				}
-			} else {
-				// Non-cancellation error (e.g. transient RPC failure) — surface it
-				// rather than silently retrying a potentially broken activity.
-				return HeartBeatDetails{}, err
 			}
+		} else if err != nil {
+			// Non-cancellation error (e.g. transient RPC failure) — surface it
+			// rather than silently retrying a potentially broken activity.
+			return HeartBeatDetails{}, err
 		}
 	}
 }
@@ -164,10 +138,9 @@ func BatchWorkflowV2(ctx workflow.Context, batchParams BatchParams) (HeartBeatDe
 // batchActivityV2 is the V2 activity for processing batch operations.
 // Compared to V1 (BatchActivity), it:
 //   - Accepts progress from a prior cancelled activity via batchParamsV2.Progress
-//   - Uses context.WithoutCancel so in-flight operations complete on cancellation
 //   - Returns current HeartBeatDetails on scan errors for resumability
 //   - Returns CanceledError with progress when context is cancelled
-func batchActivityV2(ctx context.Context, params batchParamsV2) (HeartBeatDetails, error) {
+func batchActivityV2(ctx context.Context, params BatchParams) (HeartBeatDetails, error) {
 	batcher := ctx.Value(BatcherContextKey).(*Batcher)
 	client := batcher.clientBean.GetFrontendClient()
 	var adminClient admin.Client
@@ -191,10 +164,10 @@ func batchActivityV2(ctx context.Context, params batchParamsV2) (HeartBeatDetail
 	}
 	domainID := domainResp.GetDomainInfo().GetUUID()
 
-	hbd := getHeartBeatDetailsV2(ctx)
-	// Workflow-provided progress (from a prior cancelled activity) takes precedence
-	// over heartbeat recovery, which only works within the same activity ID.
-	if params.Progress != nil {
+	hbd, ok := getHeartBeatDetails(ctx)
+	// Only fall back to workflow-provided progress when heartbeat
+	// recovery has no data (first attempt of a re-launched activity).
+	if !ok && params.Progress != nil {
 		hbd = *params.Progress
 	}
 
@@ -213,7 +186,7 @@ func batchActivityV2(ctx context.Context, params batchParamsV2) (HeartBeatDetail
 	taskCh := make(chan taskDetail, params.PageSize)
 	respCh := make(chan error, params.PageSize)
 	for i := 0; i < params.Concurrency; i++ {
-		go startTaskProcessorV2(ctx, params.BatchParams, domainID, taskCh, respCh, rateLimiter, client, adminClient)
+		go startTaskProcessor(ctx, params, domainID, taskCh, respCh, rateLimiter, client, adminClient, BatchWFV2TypeName)
 	}
 
 	for {
@@ -227,38 +200,39 @@ func batchActivityV2(ctx context.Context, params batchParamsV2) (HeartBeatDetail
 			return hbd, err
 		}
 		batchCount := len(resp.Executions)
-
-		if batchCount > 0 {
-			for _, wf := range resp.Executions {
-				taskCh <- taskDetail{
-					execution: *wf.Execution,
-					attempts:  0,
-					hbd:       hbd,
-				}
-			}
-
-			succCount := 0
-			errCount := 0
-		Loop:
-			for {
-				select {
-				case err := <-respCh:
-					if err == nil {
-						succCount++
-					} else {
-						errCount++
-					}
-					if succCount+errCount == batchCount {
-						break Loop
-					}
-				case <-ctx.Done():
-					return hbd, cadence.NewCanceledError(hbd)
-				}
-			}
-			hbd.SuccessCount += succCount
-			hbd.ErrorCount += errCount
-			hbd.CurrentPage++
+		if batchCount <= 0 {
+			break
 		}
+
+		for _, wf := range resp.Executions {
+			taskCh <- taskDetail{
+				execution: *wf.Execution,
+				attempts:  0,
+				hbd:       hbd,
+			}
+		}
+
+		succCount := 0
+		errCount := 0
+	Loop:
+		for {
+			select {
+			case err := <-respCh:
+				if err == nil {
+					succCount++
+				} else {
+					errCount++
+				}
+				if succCount+errCount == batchCount {
+					break Loop
+				}
+			case <-ctx.Done():
+				return hbd, cadence.NewCanceledError(hbd)
+			}
+		}
+		hbd.SuccessCount += succCount
+		hbd.ErrorCount += errCount
+		hbd.CurrentPage++
 		hbd.PageToken = resp.NextPageToken
 		activity.RecordHeartbeat(ctx, hbd)
 
@@ -270,114 +244,6 @@ func batchActivityV2(ctx context.Context, params batchParamsV2) (HeartBeatDetail
 			return hbd, nil
 		}
 	}
-}
 
-// getHeartBeatDetailsV2 recovers progress from a previous heartbeat, if any.
-func getHeartBeatDetailsV2(ctx context.Context) HeartBeatDetails {
-	var hbd HeartBeatDetails
-	if activity.HasHeartbeatDetails(ctx) {
-		if err := activity.GetHeartbeatDetails(ctx, &hbd); err == nil {
-			return hbd
-		}
-	}
-	return hbd
-}
-
-func startTaskProcessorV2(
-	ctx context.Context,
-	batchParams BatchParams,
-	domainID string,
-	taskCh chan taskDetail,
-	respCh chan error,
-	limiter *rate.Limiter,
-	client frontend.Client,
-	adminClient admin.Client,
-) {
-	batcher := ctx.Value(BatcherContextKey).(*Batcher)
-	// opCtx inherits all context values (batcher, activity heartbeat machinery)
-	// but is NOT cancelled when the activity context is cancelled. This allows
-	// in-flight operations to complete when a tune signal cancels the activity,
-	// so no workflows are skipped mid-page.
-	opCtx := context.WithoutCancel(ctx)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case task := <-taskCh:
-			var err error
-			requestID := uuid.New().String()
-
-			switch batchParams.BatchType {
-			case BatchTypeTerminate:
-				err = processTask(opCtx, limiter, task, batchParams, client,
-					batchParams.TerminateParams.TerminateChildren,
-					func(workflowID, runID string) error {
-						return client.TerminateWorkflowExecution(opCtx, &types.TerminateWorkflowExecutionRequest{
-							Domain: batchParams.DomainName,
-							WorkflowExecution: &types.WorkflowExecution{
-								WorkflowID: workflowID,
-								RunID:      runID,
-							},
-							Reason:   batchParams.Reason,
-							Identity: BatchWFV2TypeName,
-						})
-					})
-			case BatchTypeCancel:
-				err = processTask(opCtx, limiter, task, batchParams, client,
-					batchParams.CancelParams.CancelChildren,
-					func(workflowID, runID string) error {
-						return client.RequestCancelWorkflowExecution(opCtx, &types.RequestCancelWorkflowExecutionRequest{
-							Domain: batchParams.DomainName,
-							WorkflowExecution: &types.WorkflowExecution{
-								WorkflowID: workflowID,
-								RunID:      runID,
-							},
-							Identity:  BatchWFV2TypeName,
-							RequestID: requestID,
-						})
-					})
-			case BatchTypeSignal:
-				err = processTask(opCtx, limiter, task, batchParams, client, common.BoolPtr(false),
-					func(workflowID, runID string) error {
-						return client.SignalWorkflowExecution(opCtx, &types.SignalWorkflowExecutionRequest{
-							Domain: batchParams.DomainName,
-							WorkflowExecution: &types.WorkflowExecution{
-								WorkflowID: workflowID,
-								RunID:      runID,
-							},
-							Identity:   BatchWFV2TypeName,
-							RequestID:  requestID,
-							SignalName: batchParams.SignalParams.SignalName,
-							Input:      []byte(batchParams.SignalParams.Input),
-						})
-					})
-			case BatchTypeReplicate:
-				err = processTask(opCtx, limiter, task, batchParams, client, common.BoolPtr(false),
-					func(workflowID, runID string) error {
-						return adminClient.ResendReplicationTasks(opCtx, &types.ResendReplicationTasksRequest{
-							DomainID:      domainID,
-							WorkflowID:    workflowID,
-							RunID:         runID,
-							RemoteCluster: batchParams.ReplicateParams.SourceCluster,
-						})
-					})
-			}
-			if err != nil {
-				batcher.metricsClient.IncCounter(metrics.BatcherScope, metrics.BatcherProcessorFailures)
-				getActivityLogger(ctx).Error("Failed to process batch operation task", tag.Error(err))
-
-				_, ok := batchParams._nonRetryableErrors[err.Error()]
-				if ok || task.attempts >= batchParams.AttemptsOnRetryableError {
-					respCh <- err
-				} else {
-					// put back to the channel if less than attemptsOnError
-					task.attempts++
-					taskCh <- task
-				}
-			} else {
-				batcher.metricsClient.IncCounter(metrics.BatcherScope, metrics.BatcherProcessorSuccess)
-				respCh <- nil
-			}
-		}
-	}
+	return hbd, nil
 }
