+++ service/worker/batcher/workflow_v2_test.go
@@ -23,112 +23,113 @@
 package batcher
 
 import (
-  "context"
-  "testing"
-  "time"
+	"context"
+	"sync"
+	"testing"
+	"time"
 
-  "github.com/opentracing/opentracing-go"
-  "github.com/stretchr/testify/assert"
-  "github.com/stretchr/testify/mock"
-  "github.com/stretchr/testify/require"
-  "github.com/uber-go/tally"
-  "go.uber.org/cadence"
-  "go.uber.org/cadence/testsuite"
-  "go.uber.org/cadence/worker"
-  "go.uber.org/mock/gomock"
+	"github.com/opentracing/opentracing-go"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+	"github.com/uber-go/tally"
+	"go.uber.org/cadence"
+	"go.uber.org/cadence/testsuite"
+	"go.uber.org/cadence/worker"
+	"go.uber.org/mock/gomock"
 
-  "github.com/uber/cadence/common/metrics"
-  mmocks "github.com/uber/cadence/common/metrics/mocks"
-  "github.com/uber/cadence/common/types"
+	"github.com/uber/cadence/common/metrics"
+	mmocks "github.com/uber/cadence/common/metrics/mocks"
+	"github.com/uber/cadence/common/types"
 )
 
 func TestBatchWorkflowV2(t *testing.T) {
-  tests := []struct {
-    name        string
-    params      BatchParams
-    setupEnv    func(env *testsuite.TestWorkflowEnvironment)
-    wantErr     string
-    checkResult func(t *testing.T, result HeartBeatDetails)
-  }{
-    {
-      name:   "normal completion",
-      params: createParams(BatchTypeCancel),
-      setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
-        env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
-          Return(HeartBeatDetails{SuccessCount: 5, CurrentPage: 2}, nil)
-      },
-      checkResult: func(t *testing.T, result HeartBeatDetails) {
-        assert.Equal(t, 5, result.SuccessCount)
-        assert.Equal(t, 2, result.CurrentPage)
-      },
-    },
-    {
-      name:    "validation error on empty params",
-      params:  BatchParams{},
-      wantErr: "must provide required parameters",
-    },
-    {
-      name: "validation error missing signal name",
-      params: func() BatchParams {
-        p := createParams(BatchTypeSignal)
-        p.SignalParams.SignalName = ""
-        return p
-      }(),
-      wantErr: "must provide signal name",
-    },
-    {
-      name:   "activity returns non-retriable error",
-      params: createParams(BatchTypeTerminate),
-      setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
-        env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
-          Return(HeartBeatDetails{}, cadence.NewCustomError(_nonRetriableReason, "details"))
-      },
-      wantErr: _nonRetriableReason,
-    },
-    {
-      name:   "activity returns full result",
-      params: createParams(BatchTypeTerminate),
-      setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
-        env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
-          Return(HeartBeatDetails{SuccessCount: 10, ErrorCount: 2, CurrentPage: 3, TotalEstimate: 12}, nil)
-      },
-      checkResult: func(t *testing.T, result HeartBeatDetails) {
-        assert.Equal(t, 10, result.SuccessCount)
-        assert.Equal(t, 2, result.ErrorCount)
-        assert.Equal(t, 3, result.CurrentPage)
-        assert.Equal(t, int64(12), result.TotalEstimate)
-      },
-    },
-  }
+	tests := []struct {
+		name        string
+		params      BatchParams
+		setupEnv    func(env *testsuite.TestWorkflowEnvironment)
+		wantErr     string
+		checkResult func(t *testing.T, result HeartBeatDetails)
+	}{
+		{
+			name:   "normal completion",
+			params: createParams(BatchTypeCancel),
+			setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
+				env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
+					Return(HeartBeatDetails{SuccessCount: 5, CurrentPage: 2}, nil)
+			},
+			checkResult: func(t *testing.T, result HeartBeatDetails) {
+				assert.Equal(t, 5, result.SuccessCount)
+				assert.Equal(t, 2, result.CurrentPage)
+			},
+		},
+		{
+			name:    "validation error on empty params",
+			params:  BatchParams{},
+			wantErr: "must provide required parameters",
+		},
+		{
+			name: "validation error missing signal name",
+			params: func() BatchParams {
+				p := createParams(BatchTypeSignal)
+				p.SignalParams.SignalName = ""
+				return p
+			}(),
+			wantErr: "must provide signal name",
+		},
+		{
+			name:   "activity returns non-retriable error",
+			params: createParams(BatchTypeTerminate),
+			setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
+				env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
+					Return(HeartBeatDetails{}, cadence.NewCustomError(_nonRetriableReason, "details"))
+			},
+			wantErr: _nonRetriableReason,
+		},
+		{
+			name:   "activity returns full result",
+			params: createParams(BatchTypeTerminate),
+			setupEnv: func(env *testsuite.TestWorkflowEnvironment) {
+				env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
+					Return(HeartBeatDetails{SuccessCount: 10, ErrorCount: 2, CurrentPage: 3, TotalEstimate: 12}, nil)
+			},
+			checkResult: func(t *testing.T, result HeartBeatDetails) {
+				assert.Equal(t, 10, result.SuccessCount)
+				assert.Equal(t, 2, result.ErrorCount)
+				assert.Equal(t, 3, result.CurrentPage)
+				assert.Equal(t, int64(12), result.TotalEstimate)
+			},
+		},
+	}
 
-  for _, tt := range tests {
-    t.Run(tt.name, func(t *testing.T) {
-      var suite testsuite.WorkflowTestSuite
-      env := suite.NewTestWorkflowEnvironment()
-      env.RegisterWorkflow(BatchWorkflowV2)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var suite testsuite.WorkflowTestSuite
+			env := suite.NewTestWorkflowEnvironment()
+			env.RegisterWorkflow(BatchWorkflowV2)
 
-      if tt.setupEnv != nil {
-        tt.setupEnv(env)
-      }
+			if tt.setupEnv != nil {
+				tt.setupEnv(env)
+			}
 
-      env.ExecuteWorkflow(BatchWorkflowV2, tt.params)
-      require.True(t, env.IsWorkflowCompleted())
+			env.ExecuteWorkflow(BatchWorkflowV2, tt.params)
+			require.True(t, env.IsWorkflowCompleted())
 
-      err := env.GetWorkflowError()
-      if tt.wantErr != "" {
-        require.Error(t, err)
-        assert.Contains(t, err.Error(), tt.wantErr)
-        return
-      }
-      require.NoError(t, err)
+			err := env.GetWorkflowError()
+			if tt.wantErr != "" {
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.wantErr)
+				return
+			}
+			require.NoError(t, err)
 
-      if tt.checkResult != nil {
-        var result HeartBeatDetails
-        require.NoError(t, env.GetWorkflowResult(&result))
-        tt.checkResult(t, result)
-      }
-    })
-  }
+			if tt.checkResult != nil {
+				var result HeartBeatDetails
+				require.NoError(t, env.GetWorkflowResult(&result))
+				tt.checkResult(t, result)
+			}
+		})
+	}
 }
 
 // TestBatchWorkflowV2_TuneSignal verifies the tune-signal restart path:
@@ -141,139 +142,141 @@ func TestBatchWorkflowV2(t *testing.T) {
 // What can be verified is that the signal is received, the activity is
 // restarted, and the updated params are applied.
 func TestBatchWorkflowV2_TuneSignal(t *testing.T) {
-  var suite testsuite.WorkflowTestSuite
-  env := suite.NewTestWorkflowEnvironment()
-  env.RegisterWorkflow(BatchWorkflowV2)
+	var suite testsuite.WorkflowTestSuite
+	env := suite.NewTestWorkflowEnvironment()
+	env.RegisterWorkflow(BatchWorkflowV2)
 
-  // Track params received by each activity invocation.
-  var capturedParams []batchParamsV2
-  env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
-    Return(func(_ context.Context, p batchParamsV2) (HeartBeatDetails, error) {
-      capturedParams = append(capturedParams, p)
-      if len(capturedParams) == 1 {
-        // First invocation: the test framework will cancel this activity when
-        // the tune signal fires. Return value is discarded on cancellation;
-        // the framework delivers an empty CanceledError to the workflow.
-        return HeartBeatDetails{}, nil
-      }
-      // Second invocation: completes normally.
-      return HeartBeatDetails{SuccessCount: 8, CurrentPage: 3}, nil
-    })
+	// Track params received by each activity invocation.
+	var mu sync.Mutex
+	var capturedParams []BatchParams
+	env.OnActivity(batchActivityV2Name, mock.Anything, mock.Anything).
+		Return(func(_ context.Context, p BatchParams) (HeartBeatDetails, error) {
+			mu.Lock()
+			capturedParams = append(capturedParams, p)
+			n := len(capturedParams)
+			mu.Unlock()
+			if n == 1 {
+				// First invocation: the test framework will cancel this activity when
+				// the tune signal fires. Return value is discarded on cancellation;
+				// the framework delivers an empty CanceledError to the workflow.
+				return HeartBeatDetails{}, nil
+			}
+			// Second invocation: completes normally.
+			return HeartBeatDetails{SuccessCount: 8, CurrentPage: 3}, nil
+		})
 
-  // RegisterDelayedCallback at time.Millisecond*0 fires before any mock activity
-  // resolves, injecting the signal into the workflow's signal channel.
-  env.RegisterDelayedCallback(func() {
-    env.SignalWorkflow(SignalNameTune, TuneSignal{RPS: 20, Concurrency: 5})
-  }, time.Millisecond*0)
+	// RegisterDelayedCallback at time.Millisecond*0 fires before any mock activity
+	// resolves, injecting the signal into the workflow's signal channel.
+	env.RegisterDelayedCallback(func() {
+		env.SignalWorkflow(SignalNameTune, TuneSignal{RPS: 20, Concurrency: 5})
+	}, time.Millisecond*0)
 
-  params := createParams(BatchTypeCancel)
-  env.ExecuteWorkflow(BatchWorkflowV2, params)
+	params := createParams(BatchTypeCancel)
+	env.ExecuteWorkflow(BatchWorkflowV2, params)
 
-  require.True(t, env.IsWorkflowCompleted())
-  require.NoError(t, env.GetWorkflowError())
+	require.True(t, env.IsWorkflowCompleted())
+	require.NoError(t, env.GetWorkflowError())
 
-  var result HeartBeatDetails
-  require.NoError(t, env.GetWorkflowResult(&result))
-  assert.Equal(t, 8, result.SuccessCount)
-  assert.Equal(t, 3, result.CurrentPage)
+	var result HeartBeatDetails
+	require.NoError(t, env.GetWorkflowResult(&result))
+	assert.Equal(t, 8, result.SuccessCount)
+	assert.Equal(t, 3, result.CurrentPage)
 
-  require.Len(t, capturedParams, 2, "activity must be invoked twice (cancelled then restarted)")
-  // Second invocation must carry the updated RPS and Concurrency from the signal.
-  assert.Equal(t, 20, capturedParams[1].RPS, "RPS must be updated by tune signal")
-  assert.Equal(t, 5, capturedParams[1].Concurrency, "Concurrency must be updated by tune signal")
+	require.Len(t, capturedParams, 2, "activity must be invoked twice (cancelled then restarted)")
+	// Second invocation must carry the updated RPS and Concurrency from the signal.
+	assert.Equal(t, 20, capturedParams[1].RPS, "RPS must be updated by tune signal")
+	assert.Equal(t, 5, capturedParams[1].Concurrency, "Concurrency must be updated by tune signal")
 }
 
 func TestBatchActivityV2_UsesProgress(t *testing.T) {
-  var env testsuite.WorkflowTestSuite
-  activityEnv := env.NewTestActivityEnvironment()
-  activityEnv.RegisterActivity(batchActivityV2)
+	var env testsuite.WorkflowTestSuite
+	activityEnv := env.NewTestActivityEnvironment()
+	activityEnv.RegisterActivity(batchActivityV2)
 
-  batcher, mockResource := setuptest(t)
-  metricsMock := &mmocks.Client{}
-  metricsMock.On("IncCounter", metrics.BatcherScope, metrics.BatcherProcessorSuccess).Times(1)
-  batcher.metricsClient = metricsMock
+	batcher, mockResource := setuptest(t)
+	metricsMock := &mmocks.Client{}
+	metricsMock.On("IncCounter", metrics.BatcherScope, metrics.BatcherProcessorSuccess).Times(1)
+	batcher.metricsClient = metricsMock
 
-  mockResource.FrontendClient.EXPECT().DescribeDomain(gomock.Any(), gomock.Any()).
-    Return(&types.DescribeDomainResponse{}, nil)
-  mockResource.FrontendClient.EXPECT().RequestCancelWorkflowExecution(gomock.Any(), gomock.Any()).
-    Return(nil).AnyTimes()
-  mockResource.FrontendClient.EXPECT().DescribeWorkflowExecution(gomock.Any(), gomock.Any()).
-    Return(&types.DescribeWorkflowExecutionResponse{}, nil).AnyTimes()
+	mockResource.FrontendClient.EXPECT().DescribeDomain(gomock.Any(), gomock.Any()).
+		Return(&types.DescribeDomainResponse{}, nil)
+	mockResource.FrontendClient.EXPECT().RequestCancelWorkflowExecution(gomock.Any(), gomock.Any()).
+		Return(nil).AnyTimes()
+	mockResource.FrontendClient.EXPECT().DescribeWorkflowExecution(gomock.Any(), gomock.Any()).
+		Return(&types.DescribeWorkflowExecutionResponse{}, nil).AnyTimes()
 
-  // CountWorkflowExecutions should NOT be called because TotalEstimate > 0 via Progress.
+	// CountWorkflowExecutions should NOT be called because TotalEstimate > 0 via Progress.
 
-  // Scan with the saved token returns 1 workflow and no next page.
-  mockResource.FrontendClient.EXPECT().ScanWorkflowExecutions(gomock.Any(), &types.ListWorkflowExecutionsRequest{
-    Domain:        "test-domain",
-    PageSize:      int32(10),
-    NextPageToken: []byte("saved-token"),
-    Query:         "Closetime=missing",
-  }).Return(&types.ListWorkflowExecutionsResponse{
-    Executions:    []*types.WorkflowExecutionInfo{{Execution: &types.WorkflowExecution{WorkflowID: "wid3", RunID: "rid3"}}},
-    NextPageToken: nil,
-  }, nil)
+	// Scan with the saved token returns 1 workflow and no next page.
+	mockResource.FrontendClient.EXPECT().ScanWorkflowExecutions(gomock.Any(), &types.ListWorkflowExecutionsRequest{
+		Domain:        "test-domain",
+		PageSize:      int32(10),
+		NextPageToken: []byte("saved-token"),
+		Query:         "Closetime=missing",
+	}).Return(&types.ListWorkflowExecutionsResponse{
+		Executions:    []*types.WorkflowExecutionInfo{{Execution: &types.WorkflowExecution{WorkflowID: "wid3", RunID: "rid3"}}},
+		NextPageToken: nil,
+	}, nil)
 
-  ctx := context.WithValue(context.Background(), BatcherContextKey, batcher)
-  activityEnv.SetWorkerOptions(worker.Options{
-    MetricsScope:              tally.TestScope(nil),
-    BackgroundActivityContext: ctx,
-    Tracer:                    opentracing.GlobalTracer(),
-  })
+	ctx := context.WithValue(context.Background(), BatcherContextKey, batcher)
+	activityEnv.SetWorkerOptions(worker.Options{
+		MetricsScope:              tally.TestScope(nil),
+		BackgroundActivityContext: ctx,
+		Tracer:                    opentracing.GlobalTracer(),
+	})
 
-  params := batchParamsV2{
-    BatchParams: createParams(BatchTypeCancel),
-    Progress: &HeartBeatDetails{
-      PageToken:     []byte("saved-token"),
-      CurrentPage:   1,
-      SuccessCount:  2,
-      TotalEstimate: 4,
-    },
-  }
+	params := createParams(BatchTypeCancel)
+	params.Progress = &HeartBeatDetails{
+		PageToken:     []byte("saved-token"),
+		CurrentPage:   1,
+		SuccessCount:  2,
+		TotalEstimate: 4,
+	}
 
-  val, err := activityEnv.ExecuteActivity(batchActivityV2, params)
-  require.NoError(t, err)
+	val, err := activityEnv.ExecuteActivity(batchActivityV2, params)
+	require.NoError(t, err)
 
-  var result HeartBeatDetails
-  require.NoError(t, val.Get(&result))
-  assert.Equal(t, 3, result.SuccessCount, "should add new success to progress")
-  assert.Equal(t, 2, result.CurrentPage, "should increment from progress page")
-  assert.Equal(t, int64(4), result.TotalEstimate, "should preserve TotalEstimate from progress")
+	var result HeartBeatDetails
+	require.NoError(t, val.Get(&result))
+	assert.Equal(t, 3, result.SuccessCount, "should add new success to progress")
+	assert.Equal(t, 2, result.CurrentPage, "should increment from progress page")
+	assert.Equal(t, int64(4), result.TotalEstimate, "should preserve TotalEstimate from progress")
 }
 
 func TestBatchActivityV2_EmptyPageDoesNotIncrementCurrentPage(t *testing.T) {
-  var env testsuite.WorkflowTestSuite
-  activityEnv := env.NewTestActivityEnvironment()
-  activityEnv.RegisterActivity(batchActivityV2)
+	var env testsuite.WorkflowTestSuite
+	activityEnv := env.NewTestActivityEnvironment()
+	activityEnv.RegisterActivity(batchActivityV2)
 
-  batcher, mockResource := setuptest(t)
-  metricsMock := &mmocks.Client{}
-  batcher.metricsClient = metricsMock
+	batcher, mockResource := setuptest(t)
+	metricsMock := &mmocks.Client{}
+	batcher.metricsClient = metricsMock
 
-  mockResource.FrontendClient.EXPECT().DescribeDomain(gomock.Any(), gomock.Any()).
-    Return(&types.DescribeDomainResponse{}, nil)
-  mockResource.FrontendClient.EXPECT().CountWorkflowExecutions(gomock.Any(), gomock.Any()).
-    Return(&types.CountWorkflowExecutionsResponse{Count: 0}, nil)
+	mockResource.FrontendClient.EXPECT().DescribeDomain(gomock.Any(), gomock.Any()).
+		Return(&types.DescribeDomainResponse{}, nil)
+	mockResource.FrontendClient.EXPECT().CountWorkflowExecutions(gomock.Any(), gomock.Any()).
+		Return(&types.CountWorkflowExecutionsResponse{Count: 0}, nil)
 
-  // Scan returns no results.
-  mockResource.FrontendClient.EXPECT().ScanWorkflowExecutions(gomock.Any(), gomock.Any()).
-    Return(&types.ListWorkflowExecutionsResponse{
-      Executions:    nil,
-      NextPageToken: nil,
-    }, nil)
+	// Scan returns no results.
+	mockResource.FrontendClient.EXPECT().ScanWorkflowExecutions(gomock.Any(), gomock.Any()).
+		Return(&types.ListWorkflowExecutionsResponse{
+			Executions:    nil,
+			NextPageToken: nil,
+		}, nil)
 
-  ctx := context.WithValue(context.Background(), BatcherContextKey, batcher)
-  activityEnv.SetWorkerOptions(worker.Options{
-    MetricsScope:              tally.TestScope(nil),
-    BackgroundActivityContext: ctx,
-    Tracer:                    opentracing.GlobalTracer(),
-  })
+	ctx := context.WithValue(context.Background(), BatcherContextKey, batcher)
+	activityEnv.SetWorkerOptions(worker.Options{
+		MetricsScope:              tally.TestScope(nil),
+		BackgroundActivityContext: ctx,
+		Tracer:                    opentracing.GlobalTracer(),
+	})
 
-  params := batchParamsV2{BatchParams: createParams(BatchTypeCancel)}
-  val, err := activityEnv.ExecuteActivity(batchActivityV2, params)
-  require.NoError(t, err)
+	params := createParams(BatchTypeCancel)
+	val, err := activityEnv.ExecuteActivity(batchActivityV2, params)
+	require.NoError(t, err)
 
-  var result HeartBeatDetails
-  require.NoError(t, val.Get(&result))
-  assert.Equal(t, 0, result.CurrentPage, "empty page should not increment CurrentPage")
-  assert.Equal(t, 0, result.SuccessCount)
+	var result HeartBeatDetails
+	require.NoError(t, val.Get(&result))
+	assert.Equal(t, 0, result.CurrentPage, "empty page should not increment CurrentPage")
+	assert.Equal(t, 0, result.SuccessCount)
 }
